refactor: compose Android provider interactors in core
